# common/goods/api/api.bm.go
// Code generated by protoc-gen-bm v0.1, DO NOT EDIT.
// source: api.proto

package api

import (
	"context"

	bm "github.com/go-kratos/kratos/pkg/net/http/blademaster"
	"github.com/go-kratos/kratos/pkg/net/http/blademaster/binding"
)

// to suppressed 'imported but not used warning'
var _ *bm.Context
var _ context.Context
var _ binding.StructValidator

var PathGoodsGetGoodsDetail = "/goods"
var PathGoodsGetGoodsList = "/goods/list"
var PathGoodsGetBrandDetail = "/brand"
var PathGoodsGetBrandList = "/brand/list"

// GoodsBMServer is the server API for Goods service.
type GoodsBMServer interface {
	// 获取商品详情
	GetGoodsDetail(ctx context.Context, req *GoodsDetailReq) (resp *GoodsRes, err error)

	// 获取商品列表
	GetGoodsList(ctx context.Context, req *GoodsReq) (resp *GoodsListRes, err error)

	// 获取制造商详情
	GetBrandDetail(ctx context.Context, req *BrandsDetailReq) (resp *BrandsDetailRes, err error)

	// 获取制造商列表
	GetBrandList(ctx context.Context, req *BrandsListReq) (resp *BrandListRes, err error)
}

var GoodsSvc GoodsBMServer

func goodsGetGoodsDetail(c *bm.Context) {
	p := new(GoodsDetailReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := GoodsSvc.GetGoodsDetail(c, p)
	c.JSON(resp, err)
}

func goodsGetGoodsList(c *bm.Context) {
	p := new(GoodsReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := GoodsSvc.GetGoodsList(c, p)
	c.JSON(resp, err)
}

func goodsGetBrandDetail(c *bm.Context) {
	p := new(BrandsDetailReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := GoodsSvc.GetBrandDetail(c, p)
	c.JSON(resp, err)
}

func goodsGetBrandList(c *bm.Context) {
	p := new(BrandsListReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := GoodsSvc.GetBrandList(c, p)
	c.JSON(resp, err)
}

// RegisterGoodsBMServer Register the blademaster route
func RegisterGoodsBMServer(e *bm.Engine, server GoodsBMServer) {
	GoodsSvc = server
	e.GET("/goods", goodsGetGoodsDetail)
	e.GET("/goods/list", goodsGetGoodsList)
	e.GET("/brand", goodsGetBrandDetail)
	e.GET("/brand/list", goodsGetBrandList)
}