# common/goods/api/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/brand": {
            "get": {
                "summary": "获取制造商详情",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.goods.service.v1.BrandsDetailRes"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "description": "制造商id",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "goods.service.v1.Goods"
                ]
            }
        },
        "/brand/list": {
            "get": {
                "summary": "获取制造商列表",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.goods.service.v1.BrandListRes"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "page",
                        "description": "页数",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "size",
                        "description": "每页数量",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "goods.service.v1.Goods"
                ]
            }
        },
        "/goods": {
            "get": {
                "summary": "获取商品详情",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.goods.service.v1.GoodsRes"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "description": "商品id",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "goods.service.v1.Goods"
                ]
            }
        },
        "/goods/list": {
            "get": {
                "summary": "获取商品列表",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.goods.service.v1.GoodsListRes"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "isHot",
                        "description": "是否是热销 1:是、0:否",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "description": "页数",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "size",
                        "description": "每页数量",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "goods.service.v1.Goods"
                ]
            }
        }
    },
    "definitions": {
        ".goods.service.v1.BrandListRes": {
            "type": "object",
            "properties": {
                "brandList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/.goods.service.v1.BrandsDetailRes"
                    },
                    "description": "商品列表"
                },
                "total": {
                    "type": "integer",
                    "description": "商品总数"
                }
            },
            "description": "制造商列表Res"
        },
        ".goods.service.v1.BrandsDetailReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "制造商id"
                }
            },
            "description": "制造商详情Req"
        },
        ".goods.service.v1.BrandsDetailRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "制造商id"
                },
                "name": {
                    "type": "string",
                    "description": "制造商名称"
                },
                "simple_desc": {
                    "type": "string",
                    "description": "制造商简介"
                },
                "pic_url": {
                    "type": "string",
                    "description": "制造商图片"
                }
            },
            "description": "制造商详情Res"
        },
        ".goods.service.v1.BrandsListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "description": "页数"
                },
                "size": {
                    "type": "integer",
                    "description": "每页数量"
                }
            },
            "description": "制造商列表Req"
        },
        ".goods.service.v1.GoodsDetailReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "商品id"
                }
            },
            "description": "商品详情Req"
        },
        ".goods.service.v1.GoodsListRes": {
            "type": "object",
            "properties": {
                "goodsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/.goods.service.v1.GoodsRes"
                    },
                    "description": "商品列表"
                },
                "total": {
                    "type": "integer",
                    "description": "商品总数"
                }
            },
            "description": "商品列表Res"
        },
        ".goods.service.v1.GoodsReq": {
            "type": "object",
            "properties": {
                "isHot": {
                    "type": "integer",
                    "description": "是否是热销 1:是、0:否"
                },
                "page": {
                    "type": "integer",
                    "description": "页数"
                },
                "size": {
                    "type": "integer",
                    "description": "每页数量"
                }
            },
            "description": "商品列表Req"
        },
        ".goods.service.v1.GoodsRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "商品id"
                },
                "name": {
                    "type": "string",
                    "description": "商品名称"
                },
                "retail_price": {
                    "type": "number",
                    "format": "double",
                    "description": "零食价格"
                },
                "goods_brief": {
                    "type": "string",
                    "description": "商品简介"
                },
                "list_pic_url": {
                    "type": "string",
                    "description": "商品图片"
                }
            },
            "description": "商品详情Res"
        }
    }
}